repitbot-microservices-v2/telegram-bot/requirements.txt
python-telegram-bot[ext]>=21.0
httpx>=0.27
pydantic-settings>=2.0
